Change to use the AutoConfig library because:
The ESP32 implementation of SmartConfig can not work on any network which
is dual 5G/2.4G which is quite common now a days.  Instead of using an
Android or iOS setup app, users will be prompted to go to a little
webserver inside the device which walks them through the setup of wifi.

The current implementation of this server is fairly ugly.  It includes
a bunch of extra stuff we don't want.  For the time being users will have
to be told to go to the menu in the upper right.  See screenshots now
linked from README.md.